Made progress on FSM implementation, with some notable improvements to be made: errors need to be standardised, I don\'t know how I want it to look but some ideas include error registry; the current FSM_get_event function is bugged and will raise an error if an event hasn\'t occured.

+++ shared/inc/ringbuffer.h
@@ -11,7 +11,7 @@
 
 /*********** ringbuffer  *************/
 typedef struct ring_buf_t {
-    uint8_t* buffer;    // pointer to buffer
+    uint32_t* buffer;    // pointer to buffer
     uint32_t head;      // head index (write)
     uint32_t tail;      // tail index (read)
     uint32_t mask;      // HAS to be (2^n) - 1
@@ -26,9 +26,9 @@ typedef enum ring_buf_errors {
     RING_BUF_EMPTY
 } ring_buf_error_t;
 
-extern error_t ring_buf_setup(ring_buf_t* rb, uint8_t* buffer, uint32_t size);
+extern error_t ring_buf_setup(ring_buf_t* rb, uint32_t* buffer, uint32_t size);
 extern int ring_buf_empty(ring_buf_t* rb);
-extern error_t ring_buf_write(ring_buf_t* rb, uint8_t byte);
-extern error_t ring_buf_read(ring_buf_t* rb, uint8_t* byte);
+extern error_t ring_buf_write(ring_buf_t* rb, uint32_t byte);
+extern error_t ring_buf_read(ring_buf_t* rb, uint32_t* byte);
 
 #endif // RING_BUF_H
